requirements.txt
Django==2.2.5
django-enumfield==1.5
djangorestframework==3.10.3
mysqlclient==1.4.4
pytz==2019.2
sqlparse==0.3.0
gunicorn
django-heroku